fix: support artifact leases on Windows
